Function to get STEAM_X:X:XXXXXXXX
That bcmath can throw server fatal errors in later versions of php (+7) with the way your script implements its use. Using the function to get around this. You can also use it in openid to get the steam id while in game and go directly to the payment method selection rather than logging in.